+++ src/xscen/diagnostics.py
@@ -17,7 +17,6 @@ from xclim.core import ValidationError, dataflags
 from xclim.core.indicator import Indicator
 
 from .config import parse_config
-from .indicators import load_xclim_module
 from .utils import (
     add_attr,
     change_units,
@@ -355,11 +354,11 @@ def properties_and_measures(  # noqa: C901
     --------
     xsdba.properties : Properties module, extending from `xclim`.
     xsdba.measures : Measures modules, extending from `xclim`.
-    xclim.core.indicator.build_indicator_module_from_yaml : YAML indicator constructor function of `xclim`.
+    xclim.IndicatorCollection.from_yaml : YAML indicator constructor function of `xclim`.
     """
     if isinstance(properties, str | Path):
         logger.debug("Loading properties module.")
-        module = load_xclim_module(properties)
+        module = xc.IndicatorCollection.from_yaml(properties)
         properties = module.iter_indicators()
     elif hasattr(properties, "iter_indicators"):
         properties = properties.iter_indicators()