+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Contributors:  Juliette Lavoie (:user:`juliettelavoie`), Gabriel Rondeau-Genesse
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 * Add CanESM5-1 to IPCC_annual_global_tas.nc. (:pull:`761`).
+* Fix to work with xclim 1.0 (:pull:`764`).
+* Add ``xs.ensembles.unstack_ensembles`` function. (:pull:`764`).
 
 Bug fixes
 ^^^^^^^^^